Settings for crystallising
We want to determine the rate of crystallisation and melting at a
variety of temperatures. Important to this is finding the temperature
Tm at which the system is in equilibrium - there is no significant
change to the liquid and crystal regions.

All systems will start from the initial structure generated by lammps
and run for 8 million timesteps.