Introduced new plot - Short Order Histogram
This new plot shows how the composition of the various short range
orderings changes over time. It is envisioned that it can be one
indicator of melting or crystallising.

The plot has been incorporated into the present. There is still some
work that needs to be done prettifying the plot.