Cleanup op files generating tex files
Brought almost all the functions generating output for each type of plot
into a single function.

The format of each plot is the filename (regex type stuff works) and the
caption as a tuple. The list of all plots is located in the create.py
file and referenced in both the grouped and present each files.