Combining cool_snow and cool_tri into cool
New settings file cool which combines cool_snow and cool_tri. This
retains all the characteristics of the individual files in a single one,
all the dynamics are in a single file.

There is little bit of hackery, the trimer distance is to 1.00, while the
snowman is to 1.0. This allows the trimer and snowman to have the same
distance and different temperature gradients.