Changing defaults, contour plot
Default filetype is not png, saving on space and time to load

Radial2d now plots a contour plot for the 2d particle distribution
function. This will need some massaging of the contour lines but
otherwise should work fine. Lots of smoothing and interpolation.

Also changed frame to be angle dependent on full 2pi range, not
antiparallel.