New files to visualise the motion of the particles.
Using VMD to colour the molecules based on their number of contacts.
A movie is then exported as individual jpeg frames. These are loaded
into photoshop where the background is removed using batch.jsx. This
can the be exported to see the motion of particles and the number of
cantacts they had when they moved.